requirements.txt
flask==2.3.3
psycopg2-binary==2.9.7
pandas==2.1.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
flask-cors==4.0.0
flask-caching==2.0.2
openpyxl==3.1.2